fix(server): tapar la fuga de handles del Event Log y tres bugs silenciosos
LogThis abria un handle del Event Log con RegisterEventSource en CADA
llamada y no existia un solo DeregisterEventSource en el server: como
TraceError pasa por ahi, cada error filtraba un handle del sistema, asi
que cuanto mas fallaba el server mas se degradaba. Ahora se abre una vez y
se reusa; se libera en CerrarServidor, que es el unico camino que corre de
verdad al cerrar (el End de esa sub mata el proceso sin volver a Main). Si
ReportEvent falla se resetea el handle para reabrirlo, conservando la
auto-recuperacion que daba el registro por llamada.

Ademas:
- No se loguea mas la OpLine cruda del bridge de cuentas (llevaba el salt y
  el hash de la password) ni el email en el bridge de Patreon.
- NpcDamage declaraba Damage As Integer pero recibe GetNpcDamage, que
  devuelve Long, y lo multiplica por modificadores: un NPC de golpe alto
  desbordaba, el error 6 se lo tragaba el handler y la funcion salia
  devolviendo -1, o sea el golpe se perdia en silencio.
- Los multiplicadores de experiencia, oro y recoleccion se leian sin valor
  por defecto: una clave faltante en Configuracion.ini daba Val("") = 0 y
  la economia entera se multiplicaba por CERO sin aviso. OroPorNivel lleva
  default 0 a proposito, porque ahi el cero significa desactivado.
- HandleWalk descartaba el retorno de verifyTimeStamp en sus dos llamadas:
  si saltaba la rama de replay, CloseSocket ya habia reseteado y liberado
  el slot del usuario y el handler seguia ejecutando sobre ese slot.

Plan ia/plans/2026/agosto/08.005, ola 1.

+++ Codigo/SistemaCombate.bas
@@ -558,7 +558,10 @@ End Function
 Private Function NpcDamage(ByVal NpcIndex As Integer, ByVal UserIndex As Integer) As Long
     On Error GoTo NpcDamage_Err
     NpcDamage = -1
-    Dim Damage   As Integer, Lugar As Integer, absorbido As Integer
+    ' Damage va As Long porque GetNpcDamage devuelve Long y ademas se lo multiplica por
+    ' modificadores: con Integer, un NPC de golpe alto desbordaba, el error 6 se lo tragaba
+    ' el handler y la funcion salia devolviendo -1, o sea el golpe se perdia en silencio.
+    Dim Damage   As Long, Lugar As Integer, absorbido As Integer
     Dim defbarco As Integer
     Dim obj      As t_ObjData
     Damage = GetNpcDamage(NpcIndex)
